Added Dashing and Tuned Jumping
Controller input was not working on my computer for some reason, but
based on what I saw from Tal's code, this should work once we test this
somewhere else. Dashing can be performed by using the left and right
bumpers. Jumping is still the A button. Additionally, I increased the
negative y acceleration vector that I applied in lieu of touching
gravity.